+++ bundle/src/test/java/com/adobe/acs/commons/redirects/servlets/ImportRedirectMapServletTest.java
@@ -107,6 +107,8 @@ public class ImportRedirectMapServletTest {
         row1.getCell(3).setCellStyle(dateStyle);
         row1.createCell(4).setCellValue("note-abc");
         row1.createCell(7).setCellValue("redirects:tag1\nredirects:tag2");
+        row1.createCell(12).setCellValue(new Calendar.Builder().setDate(2025, 02, 02).build());
+        row1.getCell(12).setCellStyle(dateStyle);
 
         Row row2 = sheet.createRow(2);
         row2.createCell(0).setCellValue("/content/2");
@@ -163,6 +165,7 @@ public class ImportRedirectMapServletTest {
         assertFalse(rule3.getEvaluateURI());
         assertFalse(rule3.getContextPrefixIgnored());
         assertArrayEquals(new String[]{"redirects:tag1", "redirects:tag2"}, rule3.getTagIds());
+        assertDateEquals("02 March 2025", rule3.getEffectiveFrom());
 
         RedirectRule rule4 = rules.get("/content/2").adaptTo(RedirectRule.class);
         assertEquals("/en/we-retail", rule4.getTarget());